Move fool trap stuff to a new FoolTrap.cs
