+++ src/Patches/ItemPatches.cs
@@ -415,7 +415,7 @@ namespace TunicRandomizer {
             }
 
             if (Item.Type == ItemTypes.FOOLTRAP) {
-                (NotificationTop, NotificationBottom) = ApplyFoolEffect(itemInfo.Player);
+                (NotificationTop, NotificationBottom) = FoolTrap.ApplyFoolEffect(itemInfo.Player);
                 DisplayMessageAnyway = true;
             }
 
@@ -662,7 +662,7 @@ namespace TunicRandomizer {
             }
 
             if (Item.Type == ItemTypes.FOOLTRAP) {
-                (NotificationTop, NotificationBottom) = ApplyFoolEffect(-1);
+                (NotificationTop, NotificationBottom) = FoolTrap.ApplyFoolEffect(-1);
                 DisplayMessageAnyway = true;
             }
 
@@ -737,158 +737,6 @@ namespace TunicRandomizer {
             TunicRandomizer.Settings.SkipItemAnimations = SkipAnimationsValue;
             if (SaveFile.GetInt(GrassRandoEnabled) == 0 && TunicRandomizer.Settings.CreateSpoilerLog && !TunicRandomizer.Settings.RaceMode) {
                 ItemTracker.PopulateSpoilerLog();
-            }
-        }
-
-        public enum TrapType {
-            Ice,
-            Fire,
-            Bee,
-            Tiny,
-            Mirror,
-            Deisometric,
-        }
-
-
-        public static Dictionary<TrapType, int> TrapWeights = new Dictionary<TrapType, int> {
-            {TrapType.Ice, 30 },
-            {TrapType.Fire, 20 },
-            {TrapType.Bee, 15 },
-            {TrapType.Tiny, 15 },
-            {TrapType.Mirror, 10 },
-            {TrapType.Deisometric, 10 },
-        };
-
-
-        public static (string, string) ApplyFoolEffect(int Player, bool fromDeathLink = false) {
-            System.Random Random = new System.Random();
-            int FoolType = Random.Next(100);
-            string FoolMessageTop = $"";
-            string FoolMessageBottom = $"";
-
-            List<TrapType> weightedTrapList = new List<TrapType>();
-            foreach (TrapType trapType in TrapWeights.Keys) {
-                if (trapType == TrapType.Mirror && CameraController.Flip) {
-                    continue;
-                }
-                if (trapType == TrapType.Tiny && (PlayerCharacterPatches.TinierFox || TunicRandomizer.Settings.TinierFoxMode)) {
-                    continue;
-                }
-                if (trapType == TrapType.Bee && (PlayerCharacterPatches.StungByBee || TunicRandomizer.Settings.BiggerHeadMode)) {
-                    continue;
-                }
-                if (trapType == TrapType.Deisometric && CameraController.DerekRotationEnabled) {
-                    continue;
-                }
-                weightedTrapList.AddRange(Enumerable.Repeat(trapType, TrapWeights[trapType]));
-            }
-
-            switch (weightedTrapList[Random.Next(weightedTrapList.Count)]) {
-                case TrapType.Ice:
-                    (FoolMessageTop, FoolMessageBottom) = FoolIceTrap();
-                    break;
-                case TrapType.Fire:
-                    (FoolMessageTop, FoolMessageBottom) = FoolFireTrap();
-                    break;
-                case TrapType.Bee:
-                    (FoolMessageTop, FoolMessageBottom) = FoolBeeTrap();
-                    break;
-                case TrapType.Tiny:
-                    (FoolMessageTop, FoolMessageBottom) = FoolTinyTrap();
-                    break;
-                case TrapType.Mirror:
-                    (FoolMessageTop, FoolMessageBottom) = FoolMirrorTrap();
-                    break;
-                case TrapType.Deisometric:
-                    (FoolMessageTop, FoolMessageBottom) = FoolDeisometricTrap();
-                    break;
-            }
-
-            if (Player == -1 && IsSinglePlayer()) {
-
-            } else if (IsArchipelago() && Player != Archipelago.instance.GetPlayerSlot()) {
-                FoolMessageTop = $"\"{Archipelago.instance.GetPlayerName(Player)}\" %i^ks {FoolMessageTop}";
-            }
-            if (!fromDeathLink) {
-                Notifications.Show(FoolMessageTop, FoolMessageBottom);
-            }
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static (string, string) FoolIceTrap() {
-            PlayerCharacter.ApplyRadiationAsDamageInHP(PlayerCharacter.instance.maxhp * .2f);
-            SFX.PlayAudioClipAtFox(PlayerCharacter.instance.bigHurtSFX);
-            SFX.PlayAudioClipAtFox(PlayerCharacter.standardFreezeSFX);
-            PlayerCharacter.instance.AddFreezeTime(3f);
-            string FoolMessageTop = $"yoo R A \"<#86A5FF>FOOL<#ffffff>!!\" [fooltrap]";
-            string FoolMessageBottom = $"hahvi^ ahn Is tIm?";
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static (string, string) FoolFireTrap() {
-            PlayerCharacter.ApplyRadiationAsDamageInHP(0f);
-            PlayerCharacter.instance.stamina = 0;
-            PlayerCharacter.instance.cachedFireController.FireAmount = 3f;
-            SFX.PlayAudioClipAtFox(PlayerCharacter.instance.bigHurtSFX);
-            string FoolMessageTop = $"yoo R A \"<#FF3333>FOOL<#ffffff>!!\" [fooltrap]";
-            string FoolMessageBottom = $"iz it hawt in hEr?";
-            PlayerCharacter.instance.Flinch(true);
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static (string, string) FoolBeeTrap() {
-            SFX.PlayAudioClipAtFox(PlayerCharacter.instance.bigHurtSFX);
-            PlayerCharacter.instance.IDamageable_ReceiveDamage(PlayerCharacter.instance.hp / 3, 0, Vector3.zero, 0, 0);
-            string FoolMessageTop = $"yoo R A \"<#ffd700>FOOL<#ffffff>!!\" [fooltrap]";
-            string FoolMessageBottom = $"\"(\"it wuhz A swRm uhv <#ffd700>bEz\"...)\"";
-            PlayerCharacterPatches.StungByBee = true;
-            PlayerCharacter.instance.Flinch(true);
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static (string, string) FoolTinyTrap() {
-            SFX.PlayAudioClipAtFox(PlayerCharacter.instance.bigHurtSFX);
-            PlayerCharacter.instance.IDamageable_ReceiveDamage(PlayerCharacter.instance.hp / 3, 0, Vector3.zero, 0, 0);
-            string FoolMessageTop = $"yoo R A <#FFA500>tInE \"<#FFA500>FOOL<#ffffff>!!\" [fooltrap]";
-            string FoolMessageBottom = $"hahf #uh sIz, duhbuhl #uh kyoot.";
-            PlayerCharacterPatches.TinierFox = true;
-            PlayerCharacter.instance.Flinch(true);
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static (string, string) FoolMirrorTrap() {
-            SFX.PlayAudioClipAtFox(PlayerCharacter.instance.bigHurtSFX);
-            PlayerCharacter.instance.IDamageable_ReceiveDamage(PlayerCharacter.instance.hp / 3, 0, Vector3.zero, 0, 0);
-            string FoolMessageTop = $"[fooltrap] \"!!\"<#FF00FF>lfoo \"A ERA UOY\"";
-            string FoolMessageBottom = $"tAk uh mOmint too ruhflehkt.";
-            CameraController.Flip = true;
-            PlayerCharacter.instance.Flinch(true);
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static (string, string) FoolDeisometricTrap() {
-            System.Random Random = new System.Random();
-            SFX.PlayAudioClipAtFox(PlayerCharacter.instance.bigHurtSFX);
-            PlayerCharacter.instance.IDamageable_ReceiveDamage(PlayerCharacter.instance.hp / 3, 0, Vector3.zero, 0, 0);
-            string FoolMessageTop = $"yoo R A <#FFA500>toopointfIvdE \"<#FFA500>FOOL<#ffffff>!!\" [fooltrap]";
-            string FoolMessageBottom = $"toonik iz ehn IsOmehtrik... wAt, wuht?";
-            CameraController.DerekRotationEnabled = true;
-            List<float> cameraRotations = new List<float>() { -45f, 45f };
-            CameraController.DerekRotationRangeRight = cameraRotations[Random.Next(cameraRotations.Count)];
-            PlayerCharacter.instance.Flinch(true);
-            return (FoolMessageTop, FoolMessageBottom);
-        }
-
-        public static void CheckFoolTrapSetting(string RewardId) {
-            Reward Reward = Locations.RandomizedLocations[RewardId].Reward;
-            if (Reward.Type == "MONEY") {
-                if ((TunicRandomizer.Settings.FoolTrapIntensity == RandomizerSettings.FoolTrapOption.NORMAL && Reward.Amount < 20)
-                || (TunicRandomizer.Settings.FoolTrapIntensity == RandomizerSettings.FoolTrapOption.DOUBLE && Reward.Amount <= 20)
-                || (TunicRandomizer.Settings.FoolTrapIntensity == RandomizerSettings.FoolTrapOption.ONSLAUGHT && Reward.Amount <= 30)) {
-                    Reward.Name = "Fool";
-                    Reward.Type = "FOOL";
-                    Reward.Amount = 1;
-                }
             }
         }
